ci: use DC_SHA1=1 on osx-clang job for CI
All other jobs were using the default DC_SHA1 (which is a
recommended practice), but the default for macOS jobs being Apple's
common crypt, we didn't catch recent breakage soon enough.

We may want to give similar diversity for Linux jobs so that some of
them build with other implementations of SHA-1, but let's start
small and fill only the immediate need.

Signed-off-by: Junio C Hamano <gitster@pobox.com>